# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VoidScan Security Operations Centre API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recon/whois:
    post:
      operationId: reconWhois
      tags: [recon]
      summary: WHOIS lookup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TargetInput"
      responses:
        "200":
          description: WHOIS result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhoisResult"
  /recon/dns:
    post:
      operationId: reconDns
      tags: [recon]
      summary: DNS enumeration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TargetInput"
      responses:
        "200":
          description: DNS records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DnsResult"
  /recon/reverse-dns:
    post:
      operationId: reconReverseDns
      tags: [recon]
      summary: Reverse DNS lookup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TargetInput"
      responses:
        "200":
          description: Reverse DNS result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResult"
  /recon/asn:
    post:
      operationId: reconAsn
      tags: [recon]
      summary: ASN lookup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TargetInput"
      responses:
        "200":
          description: ASN info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AsnResult"
  /recon/geoip:
    post:
      operationId: reconGeoip
      tags: [recon]
      summary: IP Geolocation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TargetInput"
      responses:
        "200":
          description: Geolocation data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeoipResult"
  /recon/ssl:
    post:
      operationId: reconSsl
      tags: [recon]
      summary: SSL Certificate Inspector
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TargetInput"
      responses:
        "200":
          description: SSL certificate details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SslResult"
  /recon/headers:
    post:
      operationId: reconHeaders
      tags: [recon]
      summary: HTTP Header Analyzer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TargetInput"
      responses:
        "200":
          description: HTTP headers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HeadersResult"
  /recon/tech-detect:
    post:
      operationId: reconTechDetect
      tags: [recon]
      summary: Technology Detection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TargetInput"
      responses:
        "200":
          description: Detected technologies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TechDetectResult"
  /recon/subdomains:
    post:
      operationId: reconSubdomains
      tags: [recon]
      summary: Subdomain Discovery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TargetInput"
      responses:
        "200":
          description: Discovered subdomains
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubdomainResult"
  /recon/screenshot:
    post:
      operationId: reconScreenshot
      tags: [recon]
      summary: Website Screenshot Capture
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TargetInput"
      responses:
        "200":
          description: Screenshot info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScreenshotResult"

  # ─── Threat Intelligence ───────────────────────────────────────────────────
...
components:
  schemas:
...
    TargetInput:
      type: object
      required: [target]
      properties:
        target:
          type: string
...
    GenericResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        data:
          type: object
          additionalProperties: true
        error:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
    WhoisResult:
      type: object
      required: [success, target]
      properties:
        success:
          type: boolean
        target:
          type: string
        raw:
          type: ["string", "null"]
        registrar:
          type: ["string", "null"]
        createdDate:
          type: ["string", "null"]
        expiresDate:
          type: ["string", "null"]
        updatedDate:
          type: ["string", "null"]
        nameservers:
          type: array
          items:
            type: string
        status:
          type: array
          items:
            type: string
        error:
          type: ["string", "null"]
    DnsResult:
      type: object
      required: [success, target]
      properties:
        success:
          type: boolean
        target:
          type: string
        records:
          type: object
          additionalProperties: true
        error:
          type: ["string", "null"]
    AsnResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        ip:
          type: ["string", "null"]
        asn:
          type: ["string", "null"]
        org:
          type: ["string", "null"]
        isp:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        error:
          type: ["string", "null"]
    GeoipResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        ip:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        countryCode:
          type: ["string", "null"]
        region:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
        lon:
          type: ["number", "null"]
        isp:
          type: ["string", "null"]
        org:
          type: ["string", "null"]
        timezone:
          type: ["string", "null"]
        error:
          type: ["string", "null"]
    SslResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        host:
          type: ["string", "null"]
        subject:
          type: object
          additionalProperties: true
        issuer:
          type: object
          additionalProperties: true
        validFrom:
          type: ["string", "null"]
        validTo:
          type: ["string", "null"]
        daysRemaining:
          type: ["integer", "null"]
        isExpired:
          type: ["boolean", "null"]
        protocol:
          type: ["string", "null"]
        serialNumber:
          type: ["string", "null"]
        fingerprint:
          type: ["string", "null"]
        subjectAltNames:
          type: array
          items:
            type: string
        error:
          type: ["string", "null"]
    HeadersResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        url:
          type: ["string", "null"]
        statusCode:
          type: ["integer", "null"]
        headers:
          type: object
          additionalProperties: true
        securityHeaders:
          type: object
          additionalProperties: true
        missingSecurityHeaders:
          type: array
          items:
            type: string
        error:
          type: ["string", "null"]
    TechDetectResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        url:
          type: ["string", "null"]
        technologies:
          type: array
          items:
            type: object
            additionalProperties: true
        error:
          type: ["string", "null"]
    SubdomainResult:
      type: object
      required: [success, target]
      properties:
        success:
          type: boolean
        target:
          type: string
        subdomains:
          type: array
          items:
            type: object
            additionalProperties: true
        count:
          type: integer
        error:
          type: ["string", "null"]
    ScreenshotResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        url:
          type: ["string", "null"]
        screenshotUrl:
          type: ["string", "null"]
        error:
          type: ["string", "null"]
...