# lib/api-zod/src/generated/types/subdomainResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VoidScan Security Operations Centre API
 * OpenAPI spec version: 0.1.0
 */
import type { SubdomainResultSubdomainsItem } from './subdomainResultSubdomainsItem';

export interface SubdomainResult {
  success: boolean;
  target: string;
  subdomains?: SubdomainResultSubdomainsItem[];
  count?: number;
  /** @nullable */
  error?: string | null;
}